add tests for fcu->fpt  for BKE_fcurve_calc_bounds

+++ source/blender/blenkernel/intern/fcurve.c
@@ -710,8 +710,8 @@ static bool calculate_fpt_bounds(const FCurve *fcu, const float frame_range[2], 
 
   if (frame_range != NULL) {
     /* Start index can be calculated because fpt has a key on every full frame. */
-    const int start_index = (int)(frame_range[0] - fcu->fpt[0].vec[0]);
-    const int end_index = start_index + (int)(frame_range[1] - frame_range[0]);
+    const float start_index = frame_range[0] - fcu->fpt[0].vec[0];
+    const float end_index = start_index + frame_range[1] - frame_range[0];
 
     if (start_index > fcu->totvert - 1 || end_index < 0) {
       /* Range is outside of keyframe samples. */
